+++ lib/features/games/hockey/nhl/widgets/bet_button/cubit/bet_button_cubit.dart
@@ -109,8 +109,8 @@ class NhlBetButtonCubit extends Cubit<NhlBetButtonState> {
           odds: int.parse(state.mainOdds),
           betProfit: state.toWinAmount,
           gameStartDateTime: state.game.dateTime.toString(),
-          awayTeamScore: state.game.awayTeamScore,
-          homeTeamScore: state.game.homeTeamScore,
+          awayTeamScore: state.game.awayTeamScore as int,
+          homeTeamScore: state.game.homeTeamScore as int,
           uid: state.uid,
           betTeam: state.winTeam == BetButtonWin.home ? 'home' : 'away',
           dateTime: ESTDateTime.fetchTimeEST().toString(),
@@ -249,8 +249,8 @@ class NhlBetButtonCubit extends Cubit<NhlBetButtonState> {
                       betProfit: betButtonState.toWinAmount,
                       gameStartDateTime:
                           betButtonState.game.dateTime.toString(),
-                      awayTeamScore: betButtonState.game.awayTeamScore,
-                      homeTeamScore: betButtonState.game.homeTeamScore,
+                      awayTeamScore: betButtonState.game.awayTeamScore as int,
+                      homeTeamScore: betButtonState.game.homeTeamScore as int,
                       uid: currentUserId,
                       betTeam: betButtonState.winTeam == BetButtonWin.home
                           ? 'home'
